# tests/hospital-api-v1_2_openapi.yaml
openapi: 3.1.0
info:
  title: Hospital Management API
  version: 1.2.0
  description: |
    Patients, doctors, departments, OPD schedules, appointments, and the
    hospital information the call centre's voice agent reads to callers.

    This API is the SINGLE SOURCE for everything the voice agent says.
    Structured data (schedules, availability, prices, hours, rooms) is read
    live at the moment a caller asks. Prose (department descriptions,
    services, FAQ articles) is authored here and copied into the agent's
    search index by a sync job that polls `updated_since` (see "Sync").

    ## What changed in 1.2.0

    - NEW `Holidays` — hospital-wide closures. Availability returns a
      `holiday` status and booking refuses the date with `409 HOLIDAY`.
    - NEW `visit_type` on appointments (`new`, `follow_up`,
      `report_showing`) and an optional separate report-showing queue per
      session (`report_capacity`).
    - NEW `POST /schedules/import` — upload the OPD room-allocation
      spreadsheet (or CSV), preview the result, then apply it.
    - NEW `GET /audit-log` — who changed what, when, with before/after.
    - NEW `Hospital` profile — address, hotlines, emails, OPD and desk
      hours (replaces the voice agent's `CRM_OPENING_HOURS` setting).
    - NEW `Packages`, `Services` (including facility rooms) and `Articles`
      (FAQ and general content).
    - CHANGED `Department` — adds `description_bn` and `conditions`.
...

    ## What changed in 1.1.0

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - APIKeyHeader: []
...
paths:
...
  /patients:
    post:
      tags: [Patients]
      summary: Create Patient
      operationId: create_patient_patients_post
      description: |
        CHANGED in 1.1.0: accepts `Idempotency-Key`. A repeated request with
        the same key returns the original `201` body instead of creating a
        second patient.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PatientCreate' }
      responses:
        '201':
          description: Created (or the original result of a repeated Idempotency-Key)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PatientOut' }
        '409':
          description: |
            `IDEMPOTENCY_KEY_REUSED` — the key was already used with a
            different request body.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '422': { $ref: '#/components/responses/ValidationError' }
    get:
      tags: [Patients]
      summary: List Patients
      operationId: list_patients_patients_get
      parameters:
        - $ref: '#/components/parameters/Skip'
        - $ref: '#/components/parameters/Limit'
        - name: search
          in: query
          required: false
          schema: { type: [string, 'null'] }
          description: Substring match on name or mobile number.
        - name: mobile_number
          in: query
          required: false
          schema: { type: [string, 'null'], maxLength: 20 }
          description: |
            NEW in 1.1.0. EXACT match after normalising both sides to the
            11-digit national form: strip everything but digits, drop a
            leading `88`, so `+8801712345678`, `8801712345678`,
            `01712-345678` and `01712345678` all match the same patient.
            Wildcards are not interpreted.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PatientOut' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /patients/{patient_id}:
    parameters:
      - name: patient_id
        in: path
        required: true
        schema: { type: integer }
    get:
      tags: [Patients]
      summary: Get Patient
      operationId: get_patient_patients__patient_id__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PatientOut' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
    put:
      tags: [Patients]
      summary: Update Patient
      operationId: update_patient_patients__patient_id__put
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PatientUpdate' }
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PatientOut' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: [Patients]
      summary: Delete Patient
      operationId: delete_patient_patients__patient_id__delete
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }

  # ── Departments (NEW) ────────────────────────────────────────────────────
...
  /packages:
    get:
      tags: [Packages]
      summary: List Packages
      operationId: list_packages
      parameters:
        - $ref: '#/components/parameters/Skip'
        - $ref: '#/components/parameters/Limit'
        - name: search
          in: query
          required: false
          schema: { type: [string, 'null'] }
          description: Matches `name`, `name_bn`, `aliases`, and investigation names.
        - name: gender
          in: query
          required: false
          schema:
            anyOf:
              - $ref: '#/components/schemas/PackageGender'
              - type: 'null'
          description: '`male` returns `male` and `any` packages; likewise `female`.'
        - name: is_active
          in: query
          required: false
          schema: { type: [boolean, 'null'], default: true }
        - $ref: '#/components/parameters/UpdatedSince'
        - $ref: '#/components/parameters/IncludeDeleted'
      responses:
        '200':
          description: Successful Response
          headers:
            X-Server-Time: { $ref: '#/components/headers/XServerTime' }
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PackageOut' }
    post:
      tags: [Packages]
      summary: Create Package
      operationId: create_package
      description: Requires `admin` scope.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PackageCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PackageOut' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: '`DUPLICATE_NAME`'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /packages/{package_id}:
    parameters:
      - name: package_id
        in: path
        required: true
        schema: { type: integer }
    get:
      tags: [Packages]
      summary: Get Package
      operationId: get_package
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PackageOut' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Packages]
      summary: Update Package
      operationId: update_package
      description: Requires `admin` scope. Partial update.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PackageUpdate' }
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PackageOut' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: [Packages]
      summary: Delete Package
      operationId: delete_package
      description: Requires `admin` scope. Soft delete (sets `deleted_at`).
      responses:
        '204': { description: Deleted }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ── Services (NEW in 1.2.0) ──────────────────────────────────────────────
...
  /queue:
    get:
      tags: [Queue]
      summary: Today's Queue
      operationId: get_queue
      description: |
        OPTIONAL feature. The serial currently being seen in each of today's
        sessions, so the agent can answer "how far along is my serial?".
        Only meaningful if OPD desk staff update it as patients go in.
      parameters:
        - name: doctor_id
          in: query
          required: false
          schema: { type: [integer, 'null'] }
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/QueueStatus' }
  /queue/{schedule_id}:
    parameters:
      - $ref: '#/components/parameters/ScheduleId'
    put:
      tags: [Queue]
      summary: Set Now Serving
      operationId: set_now_serving
      description: |
        OPTIONAL feature. Sets today's `now_serving` for the session. Allowed
        for `admin` keys (the OPD desk). Resets automatically each day.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [now_serving]
              properties:
                now_serving: { type: integer, minimum: 0, description: '0 means not started.' }
                queue: { $ref: '#/components/schemas/QueueName' }
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QueueStatus' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ── Health (unchanged) ───────────────────────────────────────────────────
...
components:
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    Skip:
      name: skip
      in: query
      required: false
      schema: { type: integer, default: 0, minimum: 0 }
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, default: 100, minimum: 1, maximum: 500 }
...
    ScheduleId:
      name: schedule_id
      in: path
      required: true
      schema: { type: integer }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string, minLength: 8, maxLength: 128 }
      description: |
        NEW in 1.1.0. Client-generated, unique per logical operation. The
        server stores the key with a hash of the request body and the response
        for at least 24 hours. Same key + same body returns the stored
        response (same status code). Same key + different body returns
        `409 IDEMPOTENCY_KEY_REUSED`.
    UpdatedSince:
      name: updated_since
      in: query
      required: false
      schema: { type: [string, 'null'], format: date-time }
      description: |
        NEW in 1.2.0. Only records created, updated or deleted after this
        instant. Pass the previous response's `X-Server-Time`.
    IncludeDeleted:
      name: include_deleted
      in: query
      required: false
      schema: { type: boolean, default: false }
      description: |
        NEW in 1.2.0. Also return soft-deleted, inactive and unpublished
        records (each with `deleted_at` / `is_active` / `is_published` set),
        so a sync job can remove them from its index.
  headers:
    XServerTime:
      description: NEW in 1.2.0. Server time when the list was read; the next `updated_since`.
      schema: { type: string, format: date-time }
  responses:
    ValidationError:
      description: Validation Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/HTTPValidationError' }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: '`FORBIDDEN_SCOPE` — the API key''s scope does not allow this operation.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
...
    ErrorResponse:
      type: object
      description: NEW in 1.1.0. Body of every business-rule error.
      required: [code, detail]
      properties:
        code: { $ref: '#/components/schemas/ErrorCode' }
        detail:
          type: string
          description: Human-readable, English. Not for parsing.
        next_available_date:
          type: [string, 'null']
          format: date
          description: |
            On `NO_SESSION`, `SESSION_FULL` and `HOLIDAY`: the doctor's next
            date (within 31 days) with a free serial, so a client can offer
            it directly.
        holiday_name:
          type: [string, 'null']
          description: NEW in 1.2.0. On `HOLIDAY`.
        candidates:
          type: [array, 'null']
          items: { type: object }
          description: Optional supporting rows (e.g. the clashing schedule on `SCHEDULE_OVERLAP`).
    ErrorCode:
      type: string
      enum:
        - NOT_FOUND
        - PATIENT_NOT_FOUND
        - DOCTOR_NOT_FOUND
        - FORBIDDEN_SCOPE
        - PAST_DATE
        - NO_SESSION
        - SESSION_REQUIRED
        - SESSION_FULL
        - DUPLICATE_BOOKING
        - DOCTOR_NOT_ACCEPTING
        - SCHEDULE_OVERLAP
        - ROOM_CONFLICT
        - HAS_DOCTORS
        - HAS_FUTURE_APPOINTMENTS
        - DUPLICATE_NAME
        - IDEMPOTENCY_KEY_REUSED
        - HOLIDAY
        - DUPLICATE_DATE
        - IMPORT_HAS_ERRORS
        - UNSUPPORTED_FILE
...
    HTTPValidationError:
      type: object
      properties:
        detail:
          type: array
          items: { $ref: '#/components/schemas/ValidationError' }
    ValidationError:
      type: object
      required: [loc, msg, type]
      properties:
        loc:
          type: array
          items:
            anyOf: [{ type: string }, { type: integer }]
        msg: { type: string }
        type: { type: string }

    # ── patients (unchanged shape) ─────────────────────────────────────────
    PatientCreate:
      type: object
      required: [name, age, mobile_number]
      properties:
        name: { type: string, maxLength: 150 }
        age: { type: integer, minimum: 0, maximum: 150 }
        mobile_number: { type: string, maxLength: 20 }
    PatientUpdate:
      type: object
      properties:
        name: { type: [string, 'null'], maxLength: 150 }
        age: { type: [integer, 'null'], minimum: 0, maximum: 150 }
        mobile_number: { type: [string, 'null'], maxLength: 20 }
    PatientOut:
      type: object
      required: [name, age, mobile_number, id, created_at]
      properties:
        id: { type: integer }
        name: { type: string, maxLength: 150 }
        age: { type: integer, minimum: 0, maximum: 150 }
        mobile_number: { type: string, maxLength: 20 }
        created_at: { type: string, format: date-time }
        updated_at: { type: [string, 'null'], format: date-time }

    # ── departments (NEW) ──────────────────────────────────────────────────
...
    QueueName:
      type: string
      description: |
        NEW in 1.2.0. `report` only exists for sessions with
        `report_capacity > 0`.
      enum: [regular, report]
      default: regular
...
    PackageGender:
      type: string
      enum: [male, female, any]
    PackageCreate:
      type: object
      required: [name, price]
      properties:
        name: { type: string, maxLength: 150, description: 'e.g. "Gold Health Checkup (Male)"' }
        name_bn: { type: [string, 'null'], maxLength: 150 }
        aliases:
          type: array
          items: { type: string, maxLength: 80 }
          default: []
          description: 'Other names callers use: ["gold package", "গোল্ড চেকআপ"].'
        price: { type: number, minimum: 0, description: Price the caller pays, BDT. }
        regular_price:
          type: [number, 'null']
          minimum: 0
          description: Undiscounted price, BDT, when there is an offer.
        gender: { $ref: '#/components/schemas/PackageGender' }
        age_group: { type: [string, 'null'], maxLength: 80, description: 'e.g. "above 40 years"' }
        investigations:
          type: array
          items: { type: string, maxLength: 120 }
          default: []
          description: 'Tests included: ["CBC with ESR", "HbA1c", "TSH"].'
        complementary:
          type: array
          items: { type: string, maxLength: 120 }
          default: []
          description: 'Consultations and extras: ["Gynaecologist consultation"].'
        description: { type: [string, 'null'] }
        description_bn: { type: [string, 'null'] }
        preparation:
          type: [string, 'null']
          description: 'e.g. "Fasting for 10 hours before the test". Read to callers who ask.'
        is_active: { type: boolean, default: true }
    PackageUpdate:
      type: object
      description: Every field optional; omitted fields are unchanged.
      properties:
        name: { type: [string, 'null'], maxLength: 150 }
        name_bn: { type: [string, 'null'], maxLength: 150 }
        aliases: { type: [array, 'null'], items: { type: string, maxLength: 80 } }
        price: { type: [number, 'null'], minimum: 0 }
        regular_price: { type: [number, 'null'], minimum: 0 }
        gender:
          anyOf:
            - $ref: '#/components/schemas/PackageGender'
            - type: 'null'
        age_group: { type: [string, 'null'], maxLength: 80 }
        investigations: { type: [array, 'null'], items: { type: string, maxLength: 120 } }
        complementary: { type: [array, 'null'], items: { type: string, maxLength: 120 } }
        description: { type: [string, 'null'] }
        description_bn: { type: [string, 'null'] }
        preparation: { type: [string, 'null'] }
        is_active: { type: [boolean, 'null'] }
    PackageOut:
      allOf:
        - $ref: '#/components/schemas/PackageCreate'
        - type: object
          required: [id, created_at]
          properties:
            id: { type: integer }
            created_at: { type: string, format: date-time }
            updated_at: { type: [string, 'null'], format: date-time }
            deleted_at: { type: [string, 'null'], format: date-time }

    # ── services (NEW in 1.2.0) ────────────────────────────────────────────
...
    QueueStatus:
      type: object
      required: [schedule_id, doctor_id, date, queue, now_serving, booked]
      properties:
        schedule_id: { type: integer }
        doctor_id: { type: integer }
        doctor_name: { type: string }
        date: { type: string, format: date }
        queue: { $ref: '#/components/schemas/QueueName' }
        now_serving: { type: integer, description: '0 means not started.' }
        booked: { type: integer }
        updated_at: { type: [string, 'null'], format: date-time }